1024x576 is okay to show strands of hair
so i can say that the problem has been found
and i should deep research more solution for this problem.

	modified:   runtime/cpp/apps/matting/client/include/pipeline/frontend/frontend.h
	modified:   runtime/cpp/apps/matting/client/include/pipeline/inference-engine/rknn/rknn-zero-copy.h
	modified:   runtime/cpp/apps/matting/client/src/pipeline/frontend/frontend.cpp
	modified:   runtime/cpp/apps/matting/client/src/pipeline/pipeline.cpp

+++ runtime/cpp/apps/matting/client/include/pipeline/inference-engine/rknn/rknn-zero-copy.h
@@ -36,8 +36,8 @@ class InferenceEngineRKNNZeroCP : public InferenceEngine {
 	TensorData infer(const TensorData& input) override;
 
 	// Get model input dimensions (NHWC format: [batch, height, width, channels])
-	int getInputHeight() const { return static_cast<int>(input_attr_.dims[1]); }
-	int getInputWidth() const { return static_cast<int>(input_attr_.dims[2]); }
+	size_t getInputHeight() const { return static_cast<size_t>(input_attr_.dims[1]); }
+	size_t getInputWidth() const { return static_cast<size_t>(input_attr_.dims[2]); }
 
    private:
 	// member functions
